refactor: refactor global error handler
