+++ src/components/modular/index.ts
@@ -18,3 +18,4 @@ export * from './WalletSelectorItem';
 export { BalanceRow } from './BalanceRow';
 export { SingleAssetNFT } from './SingleAssetNFT';
 export { PermissionItem } from './PermissionItem';
+export { CriticalErrorHandler } from './CriticalErrorHandler';
