+++ src/entities/global-error/global-error-store.model.ts
@@ -1,0 +1,5 @@
+export interface GlobalErrorState {
+  error: { title: string; message: string } | null;
+  setError: (error: { title: string; message: string }) => void;
+  clearError: () => void;
+}
